pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "umbrela"
version = "0.0.7"
description = "A Package for generating query-passage relevance assessment labels."
readme = "README.md"
authors = [
    { name = "Shivani Upadhyay", email = "sjupadhyay@uwaterloo.ca" },
    { name = "Ronak Pradeep", email = "rpradeep@uwaterloo.ca" },
    { name = "Jimmy Lin", email = "jimmylin@uwaterloo.ca" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["LLM", "evaluation", "information retrieval"]
requires-python = ">= 3.11"
dependencies = [
    "matplotlib>=3.8.4",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.4.2",
    "tqdm>=4.66.2",
    "typing-extensions>=4.11.0",
]

[project.optional-dependencies]
cloud = [
    "google-cloud-aiplatform>=1.50.0",
    "openai>=1.24.1",
    "retry>=0.9.2",
]
hf = [
    "datasets>=2.19.0",
    "torch>=2.2.2",
    "transformers>=4.40.1",
]
fastchat = [
    "fschat>=0.2.36",
    "torch>=2.2.2",
    "transformers>=4.40.1",
]
pyserini = [
    "pyserini==1.2.0",
]
all = [
    "datasets>=2.19.0",
    "fschat>=0.2.36",
    "google-cloud-aiplatform>=1.50.0",
    "openai>=1.24.1",
    "pyserini==1.2.0",
    "retry>=0.9.2",
    "torch>=2.2.2",
    "transformers>=4.40.1",
]

[project.scripts]
umbrela-ensemble = "umbrela.ensemble_judge:main"
umbrela-gemini = "umbrela.gemini_judge:main"
umbrela-gpt = "umbrela.gpt_judge:main"
umbrela-hf = "umbrela.hgfllm_judge:main"
umbrela-os = "umbrela.osllm_judge:main"

[dependency-groups]
dev = [
    "mypy>=1.11.2",
    "pre-commit>=3.8.0",
    "pytest>=8.4.2",
    "ruff>=0.12.0",
    "types-tqdm",
]

[tool.setuptools.package-data]
umbrela = ["prompts/*.txt"]

[project.urls]
Homepage = "https://github.com/castorini/umbrela"

[tool.bumpver]
current_version = "0.0.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false


[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"', 'version = "{version}"'
]
"README.md" = [
    "{version}",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["umbrela*"]
namespaces = false

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F"]

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
cache_dir = ".mypy_cache"
incremental = true

[[tool.mypy.overrides]]
module = [
    "datasets",
    "fastchat",
    "fastchat.*",
    "pyserini",
    "pyserini.*",
    "retry",
    "sklearn",
    "sklearn.*",
    "torch",
    "torch.*",
    "transformers",
    "transformers.*",
    "vertexai",
    "vertexai.*",
]
ignore_missing_imports = true

[tool.ruff.lint.per-file-ignores]
"src/eval/test.py" = ["E501"]
